[MISC][RL] Extract shared InputSelectWrapper and styles
