+++ src/shared/dropdown-wrapper/dropdown-wrapper.styles.tsx
@@ -1,16 +1,16 @@
 import { ChevronDownIcon } from "@lifesg/react-icons/chevron-down";
 import styled, { css, keyframes } from "styled-components";
-import { Color } from "../color";
-import { DesignToken } from "../design-token";
-import { MediaQuery } from "../media";
-import { TruncateType } from "../shared/dropdown-list/types";
-import { Text, TextStyle } from "../text";
-import { Transition } from "../transition";
+import { Color } from "../../color";
+import { DesignToken } from "../../design-token";
+import { MediaQuery } from "../../media";
+import { Text, TextStyle } from "../../text";
+import { Transition } from "../../transition";
+import { TruncateType } from "../dropdown-list/types";
 
 // =============================================================================
 // STYLE INTERFACE
 // =============================================================================
-export interface InputSelectStyleProps {
+export interface DropdownWrapperStyleProps {
     disabled?: boolean;
     $readOnly?: boolean;
     error?: boolean;
@@ -76,7 +76,7 @@ const zindexPositionHide = keyframes`
 	}
 `;
 
-export const ElementBoundary = styled.div<InputSelectStyleProps>`
+export const ElementBoundary = styled.div<DropdownWrapperStyleProps>`
     position: relative;
     border: 1px solid ${Color.Neutral[5]};
     border-radius: ${BORDER_RADIUS};
@@ -144,7 +144,7 @@ export const ElementBoundary = styled.div<InputSelectStyleProps>`
     }}
 `;
 
-export const IconContainer = styled.div<InputSelectStyleProps>`
+export const IconContainer = styled.div<DropdownWrapperStyleProps>`
     transform: rotate(${(props) => (props.expanded ? 180 : 0)}deg);
     transition: ${Transition.Base};
     margin-left: 1rem;
