+++ src/shared/dropdown-wrapper/types.ts
@@ -1,0 +1,9 @@
+export interface DropdownSelectorProps {
+    children: JSX.Element[];
+    show: boolean;
+    onBlur: () => void;
+    error?: boolean | undefined;
+    disabled?: boolean | undefined;
+    testId?: string | undefined;
+    readOnly?: boolean | undefined;
+}
